requirements.txt

anthropic
pydantic
rich
google-genai